requirements.txt
streamlit
requests
pandas
fpdf
python-dotenv